feat: detect and request permissions in the popup page, close #1125
