+++ src/_locales/en/messages.json
@@ -142,6 +142,8 @@
     "person_or_group_in_list_many": "{{people}}… {{count}} in total",
     "person_or_group_in_list_many_no_preview": "{{count}} in total",
     "personas": "Personas",
+    "popup_missing_permission": "Maskbook is lacking permission to work.",
+    "popup_request_permission": "Fix it",
     "popup_current_persona": "Current Persona",
     "popup_current_wallet": "Current Wallet",
     "popup_enter_dashboard": "Enter Dashboard",
