+++ src/_locales/ja/messages.json
@@ -142,6 +142,8 @@
     "person_or_group_in_list_many": "{{people}}… 全部で{{count}}",
     "person_or_group_in_list_many_no_preview": "全部で{{count}}",
     "personas": "人格",
+    "popup_missing_permission": "",
+    "popup_request_permission": "",
     "popup_current_persona": "現在の人格アカウント",
     "popup_current_wallet": "現在の",
     "popup_enter_dashboard": "ダッシュボードに入る",
