+++ src/_locales/zh/messages.json
@@ -142,6 +142,8 @@
     "person_or_group_in_list_many": "{{people}} 等，共 {{count}} 人",
     "person_or_group_in_list_many_no_preview": "共 {{count}} 人",
     "personas": "角色",
+    "popup_missing_permission": "",
+    "popup_request_permission": "",
     "popup_current_persona": "當前角色",
     "popup_current_wallet": "",
     "popup_enter_dashboard": "進入儀表盘",
